Updates in preparation for new engine

+++ src/main/python/engine/Manager/VMManage/VBoxManage.py
@@ -1,71 +1,83 @@
-#!/usr/bin/env python
-
 from subprocess import Popen, PIPE
 import subprocess
 from sys import argv, platform
+import sys, traceback
 import logging
 import shlex
 import threading
+import sys
 from time import sleep
-from engine.VMManage.VMManage import VMManage
-from engine.VMManage.VM import VM
-import os
+from engine.Manager.VMManage.VMManage import VMManage
+from engine.Manager.VMManage.VM import VM
 import re
-import configparser
-from engine.Configuration.ConfigurationFile import ConfigurationFile
+from engine.Configuration.SystemConfigIO import SystemConfigIO
 
-class VBoxManageWin(VMManage):
+class VBoxManage(VMManage):
+
     def __init__(self):
-        logging.info("VBoxManageWin.__init__(): instantiated")
+        logging.info("VBoxManage.__init__(): instantiated")
         VMManage.__init__(self)
-        self.cf = ConfigurationFile()
-        self.vbox_path = self.cf.getConfig()['VBOX_WIN']['VBOX_PATH']
+        self.cf = SystemConfigIO()
+        self.vbox_path = self.cf.getConfig()['VBOX_LINUX']['VBOX_PATH']
 
-        self.readStatus = VMManage.MANAGER_UNKNOWN
-        self.writeStatus = VMManage.MANAGER_UNKNOWN
         #initial refresh
         #self.refreshAllVMInfo()
 
     def configureVM(self, vmName, srcIPAddress, dstIPAddress, srcPort, dstPort, adaptorNum):
         logging.info("configureVM(): instantiated")
-        #check to make sure the vm is known, if not should refresh or check name:
-        if vmName not in self.vms:
-            logging.error("configureVM(): " + vmName + " not found in list of known vms: \r\n" + str(self.vms))
-            return -1
-        t = threading.Thread(target=self.runConfigureVM, args=(vmName, srcIPAddress, dstIPAddress, srcPort, dstPort, adaptorNum))
-        t.start()
-        return 0   
+        
+        if VMManage.POSIX:
+            #check to make sure the vm is known, if not should refresh or check name:
+            if vmName not in self.vms:
+                logging.error("configureVM(): " + vmName + " not found in list of known vms: \r\n" + str(self.vms))
+                return -1
+            t = threading.Thread(target=self.runConfigureVM, args=(vmName, srcIPAddress, dstIPAddress, srcPort, dstPort, adaptorNum))
+            t.start()
+            return 0
+        else:
+            logging.error("Platform is not linux or linux2")
+            print("Sorry your platform is not supported")       
 
     def refreshAllVMInfo(self):
         logging.info("refreshAllVMInfo(): instantiated")
-
-        logging.debug("getListVMS() Starting List VMs thread")
-        t = threading.Thread(target=self.runVMSInfo)
-        t.start()
+        
+        if VMManage.POSIX:
+            logging.debug("getListVMS() Starting List VMs thread")
+            t = threading.Thread(target=self.runVMSInfo)
+            t.start()
+        else:
+            logging.error("Platform is not linux or linux2")
+            print("Sorry your platform is not supported")
         
     def refreshVMInfo(self, vmName):
         logging.info("refreshVMInfo(): instantiated: " + str(vmName))
-        logging.debug("refreshVMInfo() refresh VMs thread")
-        #check to make sure the vm is known, if not should refresh or check name:
-        if vmName not in self.vms:
-            logging.error("refreshVMInfo(): " + vmName + " not found in list of known vms: \r\n" + str(self.vms))
-            return -1
-        t = threading.Thread(target=self.runVMInfo, args=(vmName,))
-        t.start()
-        return 0
         
+        if VMManage.POSIX:
+            logging.debug("refreshVMInfo() refresh VMs thread")
+            #check to make sure the vm is known, if not should refresh or check name:
+            if vmName not in self.vms:
+                logging.error("refreshVMInfo(): " + vmName + " not found in list of known vms: \r\n" + str(self.vms))
+                return -1
+            t = threading.Thread(target=self.runVMInfo, args=(vmName,))
+            t.start()
+            return 0
+        else:
+            logging.error("Platform is not linux or linux2")
+            print("Sorry your platform is not supported")       
+    
     def runVMSInfo(self):
         logging.debug("runVMSInfo(): instantiated")
-        startupinfo = subprocess.STARTUPINFO()
-        startupinfo.dwFlags |= subprocess.STARTF_USESHOWWINDOW
         #run vboxmanage to get vm listing
         self.readStatus = VMManage.MANAGER_READING
         #clear out the current set
         self.vms = {}
         vmListCmd = self.vbox_path + " list vms"
-        logging.debug("runVMSInfo(): Collecting VM Names using cmd: " + vmListCmd)
+        if sys.platform == "linux" or sys.platform == "linux2":
+            vmListCmd = "timeout " + str(VMManage.MANAGER_STATUS_TIMEOUT_VAL) + " " + vmListCmd
+        logging.debug("runVMSInfo(): Collecting VM Names")
         try:
-            p = Popen(vmListCmd, stdout=PIPE, stderr=PIPE, startupinfo=startupinfo, encoding="utf-8")
+
+            p = Popen(shlex.split(vmListCmd, posix=self.POSIX), stdout=PIPE, stderr=PIPE, encoding="utf-8")
             while True:
                 out = p.stdout.readline()
                 if out == '' and p.poll() != None:
@@ -88,15 +100,16 @@ class VBoxManageWin(VMManage):
             vmNum = 1
             for aVM in self.vms:
                 logging.debug("runVMSInfo(): collecting # " + str(vmNum) + " of " + str(len(self.vms)))
-                vmShowInfoCmd = self.vbox_path + " showvminfo " + str(self.vms[aVM].UUID) + "" + " --machinereadable"
+                vmShowInfoCmd = self.vbox_path + " showvminfo \"" + str(self.vms[aVM].UUID) + "\"" + " --machinereadable"
+                if sys.platform == "linux" or sys.platform == "linux2":
+                    vmShowInfoCmd = "timeout " + str(VMManage.MANAGER_STATUS_TIMEOUT_VAL) + " " + vmShowInfoCmd
                 logging.debug("runVMSInfo(): Running " + vmShowInfoCmd)
-                p = Popen(vmShowInfoCmd, stdout=PIPE, stderr=PIPE, startupinfo=startupinfo, encoding="utf-8")
+                p = Popen(shlex.split(vmShowInfoCmd, posix=self.POSIX), stdout=PIPE, stderr=PIPE, encoding="utf-8")
                 while True:
                     out = p.stdout.readline()
                     if out == '' and p.poll() != None:
                         break
                     if out != '':
-                        logging.debug("runVMSInfo(): proc output: " + out)
                         #match example: nic1="none"
                         res = re.match("nic[0-9]+=", out)
                         if res:
@@ -124,18 +137,21 @@ class VBoxManageWin(VMManage):
                 vmNum = vmNum + 1
             self.readStatus = VMManage.MANAGER_IDLE
             logging.info("runVMSInfo(): Thread 2 completed: " + vmShowInfoCmd)
-        except Exception as err:
-            logging.error("Error in runVMSInfo(): " + str(err))
+        except Exception:
+            logging.error("Error in runVMSInfo(): An error occured ")
+            exc_type, exc_value, exc_traceback = sys.exc_info()
+            traceback.print_exception(exc_type, exc_value, exc_traceback)
             self.readStatus = VMManage.MANAGER_IDLE
 
     def runVMInfo(self, aVM):
         logging.debug("runVMSInfo(): instantiated")
-        startupinfo = subprocess.STARTUPINFO()
-        startupinfo.dwFlags |= subprocess.STARTF_USESHOWWINDOW
         self.readStatus = VMManage.MANAGER_READING
-        vmShowInfoCmd = self.vbox_path + " showvminfo " + self.vms[aVM].UUID + "" + " --machinereadable"
+
+        vmShowInfoCmd = self.vbox_path + " showvminfo \"" + self.vms[aVM].UUID + "\"" + " --machinereadable"
+        if sys.platform == "linux" or sys.platform == "linux2":
+            vmShowInfoCmd = "timeout " + str(VMManage.MANAGER_STATUS_TIMEOUT_VAL) + " " + vmShowInfoCmd
         logging.debug("runVMSInfo(): Running " + vmShowInfoCmd)
-        p = Popen(vmShowInfoCmd, stdout=PIPE, stderr=PIPE, startupinfo=startupinfo, encoding="utf-8")
+        p = Popen(shlex.split(vmShowInfoCmd, posix=self.POSIX), stdout=PIPE, stderr=PIPE, encoding="utf-8")
         while True:
             out = p.stdout.readline()
             if out == '' and p.poll() != None:
@@ -168,27 +184,41 @@ class VBoxManageWin(VMManage):
         logging.debug("runVMInfo(): Thread completed")
 
     def runConfigureVM(self, vmName, srcIPAddress, dstIPAddress, srcPort, dstPort, adaptorNum):
+        vmConfigVMCmd = ""
         try:
             logging.debug("runConfigureVM(): instantiated")
             self.writeStatus = VMManage.MANAGER_WRITING
             vmConfigVMCmd = self.vbox_path + " modifyvm " + str(vmName) + " --nic" + str(adaptorNum) + " generic" + " --nicgenericdrv1 UDPTunnel " + "--cableconnected" + str(adaptorNum) + " on --nicproperty" + str(adaptorNum) + " sport=" + str(srcPort) + " --nicproperty" + str(adaptorNum) + " dport=" + str(dstPort) + " --nicproperty" + str(adaptorNum) + " dest=" + str(dstIPAddress)
+            #p.wait()
+            #vmConfigVMCmd = "timeout " + str(VMManage.MANAGER_STATUS_TIMEOUT_VAL) + " " + self.vbox_path + " modifyvm " + str(vmName) + " --nic" + str(adaptorNum) + " intnet", "--intnet"+str(netNum), "TEST"
             logging.debug("runConfigureVM(): Running " + vmConfigVMCmd)
-            subprocess.check_output(vmConfigVMCmd)
-            
+            subprocess.check_output(shlex.split(vmConfigVMCmd, posix=self.POSIX))
+            #p = Popen(shlex.split(vmConfigVMCmd, posix=self.POSIX), stdout=PIPE, stderr=PIPE)
+            #while True:
+            #    out = p.stdout.readline()
+            #    if out == '' and p.poll() != None:
+            #        break
+            #    if out != '':
+            #        logging.debug("output line: " + out)
+
             self.writeStatus = VMManage.MANAGER_IDLE
             logging.debug("runConfigure(): Thread completed")
-        except Exception as err:
-            logging.error("Error: " + str(err) + " cmd: " + vmConfigVMCmd)
+        except Exception:
+            logging.error("runConfigureVM() Error: " + " cmd: " + vmConfigVMCmd)
+            exc_type, exc_value, exc_traceback = sys.exc_info()
+            traceback.print_exception(exc_type, exc_value, exc_traceback)
+            self.readStatus = VMManage.MANAGER_IDLE
+
 
     def runVMCmd(self, cmd):
         logging.debug("runVMCmd(): instantiated")
-        startupinfo = subprocess.STARTUPINFO()
-        startupinfo.dwFlags |= subprocess.STARTF_USESHOWWINDOW
         self.writeStatus = VMManage.MANAGER_WRITING
         self.readStatus = VMManage.MANAGER_READING
         vmCmd = self.vbox_path + " " + cmd
+        if sys.platform == "linux" or sys.platform == "linux2":
+            vmCmd = "timeout " + str(VMManage.MANAGER_STATUS_TIMEOUT_VAL) + " " + vmCmd
         logging.debug("runConfigureVM(): Running " + vmCmd)
-        p = Popen(vmCmd, stdout=PIPE, stderr=PIPE, startupinfo=startupinfo, encoding="utf-8")
+        p = Popen(shlex.split(vmCmd, posix=self.POSIX), stdout=PIPE, stderr=PIPE, encoding="utf-8")
         while True:
             out = p.stdout.readline()
             if out == '' and p.poll() != None:
@@ -223,42 +253,54 @@ class VBoxManageWin(VMManage):
         
     def startVM(self, vmName):
         logging.debug("startVM(): instantiated")
-        #check to make sure the vm is known, if not should refresh or check name:
-        if vmName not in self.vms:
-            logging.error("startVM(): " + vmName + " not found in list of known vms: \r\n" + str(self.vms))
-            return -1
-        cmd = "startvm " + vmName
-        t = threading.Thread(target=self.runVMCmd, args=(cmd,))
-        t.start()
-        return 0
-        
+        if VMManage.POSIX:
+            #check to make sure the vm is known, if not should refresh or check name:
+            if vmName not in self.vms:
+                logging.error("startVM(): " + vmName + " not found in list of known vms: \r\n" + str(self.vms))
+                return -1
+            cmd = "startvm " + vmName
+            t = threading.Thread(target=self.runVMCmd, args=(cmd,))
+            t.start()
+            return 0
+        else:
+            logging.error("Platform is not linux or linux2")
+            print("Sorry your platform is not supported")
+
     def suspendVM(self, vmName):
         logging.debug("suspendVM(): instantiated")
-        #check to make sure the vm is known, if not should refresh or check name:
-        if vmName not in self.vms:
-            logging.error("suspendVM(): " + vmName + " not found in list of known vms: \r\n" + str(self.vms))
-            return -1
-        cmd = "controlvm " + vmName + " savestate"
-        t = threading.Thread(target=self.runVMCmd, args=(cmd,))
-        t.start()
-        return 0
-        
+        if VMManage.POSIX:
+            #check to make sure the vm is known, if not should refresh or check name:
+            if vmName not in self.vms:
+                logging.error("suspendVM(): " + vmName + " not found in list of known vms: \r\n" + str(self.vms))
+                return -1
+            cmd = "controlvm " + vmName + " savestate"
+            t = threading.Thread(target=self.runVMCmd, args=(cmd,))
+            t.start()
+            return 0
+        else:
+            logging.error("Platform is not linux or linux2")
+            print("Sorry your platform is not supported")
+
     def stopVM(self, vmName):
         logging.debug("stopVM(): instantiated")
-        #check to make sure the vm is known, if not should refresh or check name:
-        if vmName not in self.vms:
-            logging.error("stopVM(): " + vmName + " not found in list of known vms: \r\n" + str(self.vms))
-            return -1
-        cmd = "controlvm " + vmName + " poweroff"
-        t = threading.Thread(target=self.runVMCmd, args=(cmd,))
-        t.start()
-        return 0
-        
+        if VMManage.POSIX:
+            #check to make sure the vm is known, if not should refresh or check name:
+            if vmName not in self.vms:
+                logging.error("stopVM(): " + vmName + " not found in list of known vms: \r\n" + str(self.vms))
+                return -1
+            cmd = "controlvm " + vmName + " poweroff"
+            t = threading.Thread(target=self.runVMCmd, args=(cmd,))
+            t.start()
+            return 0
+        else:
+            logging.error("Platform is not linux or linux2")
+            print("Sorry your platform is not supported")
+
 if __name__ == "__main__":
     logging.getLogger().setLevel(logging.DEBUG)
     logging.info("Starting Program")
-    logging.info("Instantiating VBoxManageWin")
-    vbm = VBoxManageWin()
+    logging.info("Instantiating VBoxManage")
+    vbm = VBoxManage()
     
     logging.info("Status without refresh: ")
     vbm.getManagerStatus()
@@ -280,54 +322,54 @@ if __name__ == "__main__":
             logging.info("adaptor: " + str(adaptor) + " Type: " + vbm.vms[vm].adaptorInfo[adaptor] + "\r\n")
     
     logging.info("Refreshing single VM Info--")
-    logging.info("Result: " + str(vbm.refreshVMInfo("\"default\"")))
+    logging.info("Result: " + str(vbm.refreshVMInfo("\"ubuntu-core4.7\"")))
 
     while vbm.getManagerStatus()["readStatus"] != VMManage.MANAGER_IDLE:
         logging.info("waiting for manager to finish query...")
         sleep(1)
     
-    logging.info("Status for \"default\"")
-    logging.info(vbm.getVMStatus("\"default\""))
+    logging.info("Status for \"ubuntu-core4.7\"")
+    logging.info(vbm.getVMStatus("\"ubuntu-core4.7\""))
 
     #runConfigureVM(self, vmName, srcIPAddress, dstIPAddress, srcPort, dstPort, adaptorNum)
-    vbm.configureVM("\"default\"", "", "127.0.0.1", 100, 100, 1)
+    vbm.configureVM("\"ubuntu-core4.7\"", "", "127.0.0.1", 100, 100, 1)
 
     while vbm.getManagerStatus()["readStatus"] != VMManage.MANAGER_IDLE and vbm.getManagerStatus()["writeStatus"] != VMManage.MANAGER_IDLE:
         logging.info("waiting for manager to finish reading/writing...")
         sleep(1)
     
-    logging.info("Result: " + str(vbm.refreshVMInfo("\"default\"")))
+    logging.info("Result: " + str(vbm.refreshVMInfo("\"ubuntu-core4.7\"")))
     while vbm.getManagerStatus()["readStatus"] != VMManage.MANAGER_IDLE and vbm.getManagerStatus()["writeStatus"] != VMManage.MANAGER_IDLE:
         logging.info("waiting for manager to finish reading/writing...")
         sleep(1)
     
-    logging.info("Status for \"default\"")
-    logging.info(vbm.getVMStatus("\"default\""))
+    logging.info("Status for \"ubuntu-core4.7\"")
+    logging.info(vbm.getVMStatus("\"ubuntu-core4.7\""))
     
     logging.info("----Testing VM commands-------")
     logging.info("----Start-------")
-    vbm.startVM("\"default\"")
+    vbm.startVM("\"ubuntu-core4.7\"")
     while vbm.getManagerStatus()["readStatus"] != VMManage.MANAGER_IDLE and vbm.getManagerStatus()["writeStatus"] != VMManage.MANAGER_IDLE:
         logging.info("waiting for manager to finish reading/writing...")
         sleep(1)
     logging.info("----Waiting 5 seconds to save state-------")
     sleep(5)
 
-    vbm.suspendVM("\"default\"")
+    vbm.suspendVM("\"ubuntu-core4.7\"")
     while vbm.getManagerStatus()["readStatus"] != VMManage.MANAGER_IDLE and vbm.getManagerStatus()["writeStatus"] != VMManage.MANAGER_IDLE:
         logging.info("waiting for manager to finish reading/writing...")
         sleep(1)
     logging.info("----Waiting 5 seconds to resume -------")
     sleep(5)
     
-    vbm.startVM("\"default\"")
+    vbm.startVM("\"ubuntu-core4.7\"")
     while vbm.getManagerStatus()["readStatus"] != VMManage.MANAGER_IDLE and vbm.getManagerStatus()["writeStatus"] != VMManage.MANAGER_IDLE:
         logging.info("waiting for manager to finish reading/writing...")
         sleep(1)
     logging.info("----Waiting 5 seconds to stop-------")
     sleep(5)
 
-    vbm.stopVM("\"default\"")
+    vbm.stopVM("\"ubuntu-core4.7\"")
     while vbm.getManagerStatus()["readStatus"] != VMManage.MANAGER_IDLE and vbm.getManagerStatus()["writeStatus"] != VMManage.MANAGER_IDLE:
         logging.info("waiting for manager to finish reading/writing...")
         sleep(1)
